queue-service/requirements.txt
fastapi
uvicorn
boto3
websockets
mysql-connector-python
pydantic